requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2==2.9.9